requirements.txt
Django==1.11.4
django-extensions==1.9.0
django-model-utils==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.6.3
gunicorn==19.7.1
py==1.4.34
pytz==2017.2
six==1.10.0
slugify==0.0.1